Separating colors and time from Printer classes.
